+++ samples/tfm_integration/psa_level_1/CMakeLists.txt
@@ -14,7 +14,6 @@ target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/psa_attestation.c)
 target_sources(app PRIVATE src/psa_crypto.c)
 target_sources(app PRIVATE src/shell.c)
-target_sources(app PRIVATE src/tfm_ipc.c)
 target_sources(app PRIVATE src/util_app_cfg.c)
 target_sources(app PRIVATE src/util_app_log.c)
 target_sources(app PRIVATE src/util_sformat.c)
